requirements.txt
flask==3.1.0
torch==2.1.2
torchvision==0.16.2
numpy==1.26.4
gunicorn==23.0.0
prometheus-flask-exporter==0.23.1
grpcio-tools
grpcio
fastapi
uvicorn[standard]
grequests
aiohttp
furl
pytest